Checkers is up and running!
The UI is a bit unresponsive at times, but the logic is fully
implemented. Ignore some of the misleading println's about moves being
available that aren't. 

All the rules are in place. If there is a jump available you need to
take it, you can make multiple jumps with the same piece in one turn,
and the game even lets you know when you've won.

Signed-off-by: Kevin Cornelius <kfc328@gmail.com>